필터 위치변경, css 적용,필요없는 컴포넌트 삭제
