+++ korea_tour/front/homestay/src/components/HouseCard.js
@@ -5,7 +5,6 @@ import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
 import CardMedia from "@material-ui/core/CardMedia";
 import CardContent from "@material-ui/core/CardContent";
-import CardActions from "@material-ui/core/CardActions";
 import IconButton from "@material-ui/core/IconButton";
 import Typography from "@material-ui/core/Typography";
 import ShareIcon from "@material-ui/icons/Share";
@@ -17,7 +16,8 @@ import FavoriteIcon from '@material-ui/icons/Favorite';
 import FavoriteBorderIcon from '@material-ui/icons/FavoriteBorder';
 import { pink } from '@material-ui/core/colors';
 import { withRouter } from 'react-router-dom';
-//import HouseListSort from './HouseListSort';
+import './HouseCard.css';
+
 
 const useStyles = makeStyles(() => ({
   root: { maxwidth: 365 },
@@ -83,21 +83,19 @@ function HouseCard(props) {
         <Typography variant="body2" color="textSecondary" component="p">
          {props.addr1}
         </Typography>
-        <Star /> {props.avgOfStar}
-         <Box component="fieldset" mb={3} borderColor="transparent">
+        <div class="starAndReview">
+         <Star />
+         <Typography component="legend" color="textSecondary"> {props.avgOfStar}({props.countOfReview})</Typography></div>
          <Typography id="rate" variant="legend" color="textSecondary" component="p">
-        {props.price}</Typography>
-         <Typography component="legend" id="reviewCnt" color="textSecondary">({props.countOfReview})</Typography>
-         <Typography component="legend" id="price" color="textSecondary"></Typography>
-        </Box>
+         ₩ {props.price}원</Typography>
       </CardContent>
-      <CardActions disableSpacing>
-      <Button color="secondary" onClick={() => {
+     
+      <div class="goDetail">
+      <Button  color="secondary" onClick={() => {
         props.history.push(
         `/homestay/housedetail/${props.homeStayNum}` )
-      }}>자세히보기>></Button>
-      </CardActions>
-      {/*<HouseListSort /> */}
+      }}>자세히보기>></Button></div>
+
     </Card>
   );
 }
