+++ korea_tour/front/homestay/src/components/HouseListSort.js
@@ -6,31 +6,6 @@ import Button from '@material-ui/core/Button';
 import { withRouter } from 'react-router-dom';
 import axios from 'axios';
 
-const byCost = async () => {
-  try {
-    
-    const response = await axios.patch(
-      `/homestays/price/1`
-    );
-    console.log("성공");
-  } catch (e) {
-    console.log(e);
-  }
-};
-
-const byStars = async () => {
-  try {
-   
-    const response = await axios.patch(
-      `/homestays/review/1`
-    );
-    console.log("성공");
-  } catch (e) {
-    console.log(e);
-  }
-};
-
-
 const useStyles = makeStyles((theme) => ({
   mainFeatured: {
     height: 180,
@@ -65,11 +40,7 @@ export default function HouseListSort(props) {
           근처 숙소들 :)
           <br/>
         </Typography>
-        
-
         <div className={classes.button}>
-        <Button variant="contained" size="small" onClick={byCost}>요금순</Button>&nbsp;&nbsp;
-        <Button variant="contained" color="secondary" size="small" id="orderByStars" onClick={byStars}>평점순</Button>
         </div>
       </Grid>
     </Grid>
